requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
spacy>=3.7.0
nltk>=3.8.1
pdfplumber>=0.10.0
python-docx>=1.1.0
